+++ frontend/src/components/Login/PrivateRouter.jsx
@@ -1,11 +1,11 @@
 import { Outlet } from 'react-router-dom';
 import { Navigate } from 'react-router-dom';
 import { loginState } from '../../recoil/login';
-import { useRecoilState } from 'recoil';
+import { useRecoilState, useRecoilValue } from 'recoil';
 
 const PrivateRouter = () => {
-  const [isAuthorized, setIsAuthorized] = useRecoilState(loginState);
-  console.log('isAuthorized', isAuthorized);
+  const isAuthorized = useRecoilValue(loginState);
+
   return (
     <>{isAuthorized ? <Outlet /> : <Navigate to='/login' rereplace={true} />}</>
   );
