+++ frontend/src/components/Login/PublicRouter.jsx
@@ -1,11 +1,11 @@
 import { Outlet } from 'react-router-dom';
 import { Navigate } from 'react-router-dom';
 import { loginState } from '../../recoil/login';
-import { useRecoilState } from 'recoil';
+import { useRecoilState, useRecoilValue } from 'recoil';
 
 const PublicRouter = () => {
-  const [isAuthorized, setIsAuthorized] = useRecoilState(loginState);
-  console.log('isAuthorized', isAuthorized);
+  const isAuthorized = useRecoilValue(loginState);
+
   return (
     <>{isAuthorized ? <Navigate to='/' rereplace={true} /> : <Outlet />} </>
   );
